Address failing tests
